backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
PyJWT
alembic
passlib[bcrypt]
python-jose 